// frontend/src/openapi/model/permisosDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TrabajadorDto } from './trabajadorDto';


export interface PermisosDto { 
    id?: number;
    permiso?: PermisosDto.PermisoEnum;
    trabajador?: TrabajadorDto;
    fecha?: string;
    hora?: number;
    estado?: PermisosDto.EstadoEnum;
}
export namespace PermisosDto {
    export type PermisoEnum = 'VACACIONES' | 'INDISPOSICION_MEDICA' | 'ASISTENCIA_MEDICA' | 'HOSPITALIZACION_FAMILIAR' | 'INTERVENCION_QUIRURGICA' | 'ENFERMEDAD' | 'ACCIDENTE_GRAVE' | 'FALLECIMIENTO_FAMILIAR' | 'MATRIMONIO' | 'EXAMEN_PRENATAL' | 'NACIMIENTO_HIJO' | 'PERMISO_MATERNIDAD' | 'PERMISO_PATERNIDAD' | 'LACTANCIA' | 'ASISTENCIA_EXAMEN' | 'MUDANZA' | 'BAJA_ACCIDENTE' | 'BAJA_ENFERMEDAD';
    export const PermisoEnum = {
        Vacaciones: 'VACACIONES' as PermisoEnum,
        IndisposicionMedica: 'INDISPOSICION_MEDICA' as PermisoEnum,
        AsistenciaMedica: 'ASISTENCIA_MEDICA' as PermisoEnum,
        HospitalizacionFamiliar: 'HOSPITALIZACION_FAMILIAR' as PermisoEnum,
        IntervencionQuirurgica: 'INTERVENCION_QUIRURGICA' as PermisoEnum,
        Enfermedad: 'ENFERMEDAD' as PermisoEnum,
        AccidenteGrave: 'ACCIDENTE_GRAVE' as PermisoEnum,
        FallecimientoFamiliar: 'FALLECIMIENTO_FAMILIAR' as PermisoEnum,
        Matrimonio: 'MATRIMONIO' as PermisoEnum,
        ExamenPrenatal: 'EXAMEN_PRENATAL' as PermisoEnum,
        NacimientoHijo: 'NACIMIENTO_HIJO' as PermisoEnum,
        PermisoMaternidad: 'PERMISO_MATERNIDAD' as PermisoEnum,
        PermisoPaternidad: 'PERMISO_PATERNIDAD' as PermisoEnum,
        Lactancia: 'LACTANCIA' as PermisoEnum,
        AsistenciaExamen: 'ASISTENCIA_EXAMEN' as PermisoEnum,
        Mudanza: 'MUDANZA' as PermisoEnum,
        BajaAccidente: 'BAJA_ACCIDENTE' as PermisoEnum,
        BajaEnfermedad: 'BAJA_ENFERMEDAD' as PermisoEnum
    };
    export type EstadoEnum = 'SOLICITADO' | 'APROBADO' | 'RECHAZADO';
    export const EstadoEnum = {
        Solicitado: 'SOLICITADO' as EstadoEnum,
        Aprobado: 'APROBADO' as EstadoEnum,
        Rechazado: 'RECHAZADO' as EstadoEnum
    };
}