// frontend/src/openapi/model/tipoPermisoDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TipoPermisoDto { 
    id?: number;
    nombre?: string;
    codigo?: string;
}